Android gets cursors and context menus, and the core had not compiled there for two days
BackendFeatures.Cursors and BackendFeatures.ContextMenus on AndroidBackend, and
one defect found by the first Android build since M10 landed.

Context menus. Android raises one two ways and keeps them on two listener slots:
a finger long-presses, a mouse or stylus secondary-clicks. Both are installed,
and both raise the PopupMenu that AttachedMenus already builds rather than a
second menu representation. registerForContextMenu was not used: it routes
through Activity.onCreateContextMenu, a callback that does not know which view
asked. The one conflict -- a view is allowed exactly one long-click listener,
and updateTapGestureTarget uses it for .onTapGesture(.longPress) -- is written
out at the site rather than left to be discovered.

Driven on emulator-5554 in two halves, because no single tool reaches both
windows. actions/android/P72-context-menu.csv long-presses the mesh view and the
capture shows a popup carrying "Reset the camera" and "Snapshot"; then
`adb shell input tap 254 1063` produced
"CONTEXT MENU reset the camera: dist 3.40 high 1.30" in logcat, which is the
item's action having run. Pressing the item cannot be a row: dispatchTouchEvent
reaches this activity's window and a PopupMenu is a separate one, the same limit
that stops P2, P17, P19 and P20 after one click. Proved it can fail: the same
file with the press moved to the title text produces no popup.

longpress is now a real hold on Android. Android does not read a duration out of
the events -- View.onTouchEvent posts a CheckForLongPress runnable on ACTION_DOWN
and cancels it on ACTION_UP -- so a down and an up posted back to back with
timestamps 500 ms apart is a tap. The sleep stays on the replay thread so the
main thread is free to run the runnable this verb exists to trigger.

Cursors: View.setPointerIcon with PointerIcon.getSystemIcon, all seven cases
mapped to real system icons. IMPLEMENTED AND NOT DRIVEN, and the reason is
stated rather than assumed: this AVD has no pointer device. dumpsys input lists
gpio-keys and twelve virtio_input_multi_touch and nothing else,
MousePointerControllers is empty, and three attempts to register a virtual mouse
through uinput (as root and as shell) produced no device. With no
PointerController there is no sprite to draw. What is verified is that the
conformance is taken: the CursorDegradation warning is gone from logcat while
the Mesh3DViews, ScrollGestures and KeyEvents warnings are still there.

The defect. Mesh3DExport.swift called sin() on a Float. Darwin's maths module
has that overload; Bionic's math.h has sin(double) and sinf(float) and nothing
in between, so on Android the compiler could not settle what `/` meant either
and reported six errors. testapp/P72.swift had the same line. That file is not
in a backend -- it is in SwiftCrossUI, the core every platform links -- so
AndroidBackend had produced no binary at all since the day it was added, while
macOS and iOS were green and reported done. Both now compute in Double and
convert, which needs no #if. Recorded as mistakes entry 24, the first occurrence
in this tree of gate 4 of the mistakes_prevention skill: fnmatch is POSIX,
sin(Float) is Darwin, same class.

After the fix, actions/mac/P72-stop-and-check.csv reproduces all ten assertions
on macOS and the exported .glb still loads in three.js's GLTFLoader (2356 bytes,
node euler z 1.0829 rad). coverage-matrix.csv2 gains the six M10 and 10.7 rows,
which P72 did not have.

+++ testapp/P72.swift
@@ -418,14 +418,27 @@ final class P72Model: SwiftCrossUI.ObservableObject {
         //
         // 自轉是兩者中較慢的那一個,好讓兩種運動用肉眼就分得開。兩者都由同一個 `FrameClocks` 跳動驅動。
         let radius = cameraDistance
-        let orbit = Float(angle)
+        // `Double` for the turn, `Float` for the result. `sin(Float)` exists on
+        // Darwin and not on Android -- Bionic's math.h has `sin(double)` and
+        // `sinf(float)` and nothing else -- so this same line built on macOS and
+        // did not type-check at all for Android. Same fix, same reason, as
+        // `quaternion(fromEulerZYX:)` in Mesh3DExport.swift.
+        // 轉角用 `Double`、結果用 `Float`。`sin(Float)` 在 Darwin 上存在,在 Android 上不存在
+        // ——Bionic 的 math.h 只有 `sin(double)` 與 `sinf(float)`,沒有別的——因此同一行在 macOS 上
+        // 建得起來,對 Android 卻連型別檢查都過不了。與 Mesh3DExport.swift 的
+        // `quaternion(fromEulerZYX:)` 是同一個修法、同一個理由。
+        let orbit = Double(angle)
         let spin = spinning ? Float(angle) * 0.45 : 0
         var cube = P72Cube.mesh
         cube.transform = .rotatedInXY(spin)
         return Mesh3DScene(
             meshes: [cube],
             camera: Mesh3DCamera(
-                position: SIMD3(radius * sin(orbit), cameraHeight, radius * cos(orbit)),
+                position: SIMD3(
+                    Float(Double(radius) * sin(orbit)),
+                    cameraHeight,
+                    Float(Double(radius) * cos(orbit))
+                ),
                 target: SIMD3(0, 0, 0),
                 fieldOfView: 45
             ),
